WIP: Add panel UI shell
