+++ package.json
@@ -20,13 +20,16 @@
     "@types/webextension-polyfill": "^0.10.0",
     "@typescript-eslint/eslint-plugin": "^5.48.2",
     "@typescript-eslint/parser": "^5.48.2",
+    "autoprefixer": "^10.4.14",
     "eslint": "^8.32.0",
     "eslint-plugin-svelte": "^2.14.1",
     "eslint-plugin-svelte3": "^4.0.0",
     "npm-run-all": "^4.1.5",
+    "postcss": "^8.4.21",
     "svelte": "^3.55.1",
     "svelte-check": "^3.0.2",
     "svelte-preprocess": "^5.0.0",
+    "tailwindcss": "^3.2.7",
     "tslib": "^2.4.1",
     "typescript": "^4.9.4",
     "vite": "^4.0.4",
