+++ tailwind.config.cjs
@@ -1,0 +1,10 @@
+/** @type {import('tailwindcss').Config} */
+module.exports = {
+  content: [
+    './src/devtools/**/*.{html,js,svelte,ts}'
+  ],
+  theme: {
+    extend: {},
+  },
+  plugins: [],
+}
